requirements.txt
apache-airflow==2.7.1
psycopg2-binary
yfinance
pandas
